added temp working example
