+++ api/src/servers/facade.rs
@@ -4,30 +4,25 @@ use protos::roomba_server::{Roomba, RoombaServer};
 use protos::{LightBumper, SensorData, SensorsReceived, SensorsRequest, Stasis};
 
 // get standard library utils
+use async_std::task;
 use std::marker::Sync;
 use std::pin::Pin;
 use std::sync::{Arc, Mutex};
+extern crate proc_macro;
+use proc_macro::TokenStream;
+
+use futures_core::stream::Stream;
+use futures_util::pin_mut;
+use futures_util::stream::StreamExt;
 
 // gRPC tools
 use crate::servers::endpoints::RoombaService;
-use futures::{Stream, StreamExt};
+// use futures::{Stream, StreamExt};
+// use proc_macro::TokenStream;
+use crate::servers::utils::SyncBoxStream;
+use std::thread;
+use tokio::sync::mpsc;
 use tonic::{Request, Response, Status};
-
-impl RoombaService {
-    pub fn push_sensor_data_to_buffer(&self, sensor_data: SensorData) {
-        let buffer_clone = self.sensor_buffer.clone();
-        buffer_clone.lock().unwrap().push(sensor_data);
-    }
-
-    pub fn pop_sensor_data_from_buffer(&self) -> Option<SensorData> {
-        let mut sensor_buffer = self.sensor_buffer.lock().unwrap();
-        if sensor_buffer.len() > 0 {
-            Some(sensor_buffer.remove(0))
-        } else {
-            None
-        }
-    }
-}
 
 #[tonic::async_trait]
 impl Roomba for RoombaService {
@@ -39,16 +34,35 @@ impl Roomba for RoombaService {
     }
 
     // define type alias
-    #[rustfmt::skip]
-    type GetSensorDataStream = Pin<Box<dyn Stream<Item = Result<SensorData, Status>>
-    + Send
-    + Sync
-    +'static >>;
+    // #[rustfmt::skip]
+    // type GetSensorDataStream = Pin<Box<dyn Stream<Item = Result<SensorData, Status>>
+    // + Send
+    // + Sync
+    // + 'static
+    // >>;
+    type GetSensorDataStream = SyncBoxStream<'static, Result<SensorData, Status>>;
 
     async fn get_sensor_data(
         &self,
         request: Request<SensorsRequest>,
     ) -> Result<Response<Self::GetSensorDataStream>, Status> {
-        unimplemented!("todo")
+        println!("request = {:?}", request);
+
+        let (tx, rx) = mpsc::channel(4);
+
+        // tokio::spawn(async {
+        //     loop {
+        //         if let Some(data) = self.pop_sensor_data_from_buffer() {
+        //             tx.send(Ok(data)).await.unwrap();
+        //         }
+        //     }
+        // });
+        if let Some(data) = self.pop_sensor_data_from_buffer() {
+            tx.send(Ok(data)).await.unwrap();
+        }
+
+        Ok(Response::new(Box::pin(
+            tokio_stream::wrappers::ReceiverStream::new(rx),
+        )))
     }
 }
